Refactor every file and add temperature.py and rain.py
